Creation of a script that generates sanitized archives
It is easy to compress a set of files including build
files and private credentials. This script checks against
a blocklist, much like other ignore files, and prevents
them from being added to the archive.

Named 'archive' in order to accomodate support for
other potential archiving formats in the future.

Bug: 116828210
Change-Id: I02f204b09640399cc7285aa131213e05b082a862